+++ src/code42cli/click_ext/types.py
@@ -147,4 +147,22 @@ class MapChoice(click.Choice):
     def convert(self, value, param, ctx):
         if value in self.extras_map:
             value = self.extras_map[value]
+
         return super().convert(value, param, ctx)
+
+
+class TOTP(click.ParamType):
+    """Validates param to be a 6-digit integer, which is what all Code42 TOTP tokens will be."""
+
+    def get_metavar(self, param):
+        return "TEXT"
+
+    def convert(self, value, param, ctx):
+        try:
+            int(value)
+            assert len(value) == 6
+            return value
+        except Exception:
+            raise BadParameter(
+                f"TOTP tokens should be a 6-digit integer. '{value}' was provided."
+            )