+++ CHANGELOG.md
@@ -8,12 +8,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
-
 ## Unreleased
 
 ### Added
 
 - New command 'code42 users list --include-legal-hold-membership' to print the legal hold matter name and ID for any user on legal hold.
+
+## 1.9.0 - 2021-08-19
+
+### Added
+
+- `code42 profile` commands that validate passwords (`create`, `update`, `reset-pw`) now have the `--debug` option available, and `create` and `update` can now also pass in `--totp` as an option.
+
+- New command options for `code42 security-data search`
+    - `--risk-indicator` to filter events by risk indicators.
+    - `--risk-severity` to filter events by risk severity.
+
+### Changed
+
+- A TOTP token is now required on `code42 profile` commands that check for password validity when a user has MFA enabled.
+
+- Updated minimum version of py42 to `1.18.0` to provide access to `FIRST_DESTINATION_USE` and `RARE_DESTINATION_USE` search filters.
+
+### Fixed
+
+- `code42 profile delete` command now prints a clear error message when deletion target doesn't exist.
 
 ## 1.8.1 - 2021-07-14